test: expand all thin test files to 50+ methods each (5989 total)
Systematically brought every backend/tests/unit/ file (except
test_benchmarks.py) to >= 50 test methods by appending focused
class-based suites. 71 new tests added this session; prior session
accumulated source + test changes are included.

Key fixes during expansion:
- EddyCurrentModel: no loss_W() -- use eddy_loss_W() signature
- PerformanceAnalyzer: no reset() -- use frequency accumulation tests
- QueryCache: no get() -- use _cache dict + .data attribute
- BackendInfo: backend_type (not backend), reason (not message)
- card_compiler: Opcode.ADD not Opcode.LOAD
- cpp/java services: no machine_code attr -- use status attr
- antikythera_gears: import from antikythera not antikythera_gears
- BabbageNumber: no __neg__ -- use subtraction-based tests
- babbage_programs: programs at docs/simulation/programs/ not tools/ae_programs/
- Metrics ExecutionContext: import from backend.src.metrics (not services.metrics)

+++ backend/src/emulator/simulation/engine.py
@@ -101,7 +101,7 @@ class SimulationEngine:
 
     def _total_gravity_load_N(self) -> float:
         """Dead weight of the machine [N]."""
-        return self.config.machine_mass_kg * 9.81
+        return float(self.config.machine_mass_kg * 9.81)
 
     # ------------------------------------------------------------------
     # Single Step
@@ -338,11 +338,11 @@ class SimulationEngine:
 
         regime = self.state.lubrication_regime
         if regime == "full_film":
-            return base_mu * 0.05  # ~95% reduction
+            return float(base_mu * 0.05)  # ~95% reduction
         elif regime == "mixed":
-            return base_mu * 0.3  # ~70% reduction
+            return float(base_mu * 0.3)  # ~70% reduction
         else:
-            return base_mu  # boundary/dry
+            return float(base_mu)  # boundary/dry
 
     def _update_lubrication(self) -> tuple[float, str]:
         """Compute lambda ratio and lubrication regime.
